Update x-ms-external test to non-polymorphic type
Simplifies the test as we don't have to copy the custom SerDe methods.
