+++ packages/autorest.go/test/maps/azalias/zz_models_serde.go
@@ -211,40 +211,32 @@ func (g *GeoJSONObject) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-// MarshalJSON implements the json.Marshaller interface for type ListItem.
-func (l ListItem) MarshalJSON() ([]byte, error) {
+// MarshalJSON implements the json.Marshaller interface for type GeoJSONObjectNamedCollection.
+func (g GeoJSONObjectNamedCollection) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
-	populate(objectMap, "aliasId", l.AliasID)
-	populate(objectMap, "createdTimestamp", l.CreatedTimestamp)
-	populate(objectMap, "creatorDataItemId", l.CreatorDataItemID)
-	populate(objectMap, "lastUpdatedTimestamp", l.LastUpdatedTimestamp)
+	populate(objectMap, "collectionName", g.CollectionName)
+	populate(objectMap, "objects", g.Objects)
 	return json.Marshal(objectMap)
 }
 
-// UnmarshalJSON implements the json.Unmarshaller interface for type ListItem.
-func (l *ListItem) UnmarshalJSON(data []byte) error {
+// UnmarshalJSON implements the json.Unmarshaller interface for type GeoJSONObjectNamedCollection.
+func (g *GeoJSONObjectNamedCollection) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", l, err)
+		return fmt.Errorf("unmarshalling type %T: %v", g, err)
 	}
 	for key, val := range rawMsg {
 		var err error
 		switch key {
-		case "aliasId":
-			err = unpopulate(val, "AliasID", &l.AliasID)
+		case "collectionName":
+			err = unpopulate(val, "CollectionName", &g.CollectionName)
 			delete(rawMsg, key)
-		case "createdTimestamp":
-			err = unpopulate(val, "CreatedTimestamp", &l.CreatedTimestamp)
-			delete(rawMsg, key)
-		case "creatorDataItemId":
-			err = unpopulate(val, "CreatorDataItemID", &l.CreatorDataItemID)
-			delete(rawMsg, key)
-		case "lastUpdatedTimestamp":
-			err = unpopulate(val, "LastUpdatedTimestamp", &l.LastUpdatedTimestamp)
+		case "objects":
+			g.Objects, err = unmarshalGeoJSONObjectClassificationMap(val)
 			delete(rawMsg, key)
 		}
 		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", l, err)
+			return fmt.Errorf("unmarshalling type %T: %v", g, err)
 		}
 	}
 	return nil
